+++ examples/flutter_mobile_full/lib/main.dart
@@ -2,7 +2,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:form_concierge_client/form_concierge_client.dart';
 import 'package:form_concierge/form_concierge.dart';
 import 'package:shared_preferences/shared_preferences.dart';