+++ examples/flutter_mobile_full/test/replies_screen_test.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_mobile_full/main.dart';
 import 'package:form_concierge/form_concierge.dart';
-import 'package:form_concierge_client/form_concierge_client.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:shared_preferences/shared_preferences.dart';