+++ widget/lib/client.dart
@@ -1,0 +1,1 @@
+export 'package:form_concierge_client/form_concierge_client.dart';